requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
humanfriendly==10.0
jax==0.7.2
jaxlib==0.7.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.14
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.19.0
onnxruntime==1.23.1
onnxruntime-tools==1.7.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==4.25.8
psutil==7.1.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pycparser==2.23
pyparsing==3.2.5
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
xmltodict==1.0.2
